Random Variables

MCIntegration.Dist.Continuous — Method

function Continuous(lower::Float64, upper::Float64; ninc = 1000, alpha=2.0, adapt=true) where {G}

Continuous real-valued variables on the domain [lower, upper). MC will learn the distribution using the Vegas algorithm and then perform an imporant sampling accordingly.

Arguments:

lower : lower bound

upper : upper bound

ninc : number of increments

alpha : learning rate

adapt : turn on or off the adaptive map

Remark:

Vegas map maps the original integration variables x into new variables y, so that the integrand is as flat as possible in y:

\[\begin{aligned} x_0 &= a \\ x_1 &= x_0 + \Delta x_0 \\ x_2 &= x_1 + \Delta x_1 \\ \cdots \\ x_N &= x_{N-1} + \Delta x_{N-1} = b \end{aligned}\]

where a and b are the limits of integration. The grid specifies the transformation function at the points $y=i/N$ for $i=0,1\ldots N$:

\[x(y=i/N) = x_i\]

Linear interpolation is used between those points. The Jacobian for this transformation is:

\[J(y) = J_i = N \Delta x_i\]

The grid point $x_i$ is trained after each iteration.

MCIntegration.Dist.Discrete — Type

function Discrete(lower::Int, upper::Int; distribution=nothing, alpha=2.0, adapt=true)

Integer variables in the closed set [lower, upper]. MC will learn the distribution and perform an imporant sampling accordingly.

MCIntegration.Dist.create! — Method

@inline function create!(T::Continuous, idx::Int, config)

Propose to generate new (uniform) variable randomly in [T.lower, T.lower+T.range), return proposal probability

Arguments

T: Continuous variable

idx: T.data[idx] will be updated

MCIntegration.Dist.create! — Method

create!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to generate new index (uniformly) randomly in [1, size]

Arguments

newIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

MCIntegration.Dist.create! — Method

create!(K::FermiK{D}, idx::Int, rng=GLOBAL_RNG)

Propose to generate new Fermi K in [Kf-δK, Kf+δK)

Arguments

newK: vector of dimension of d=2 or 3

MCIntegration.Dist.locate — Method

function locate(accumulation, p)

Return index of p in accumulation so that accumulation[idx]<=p<accumulation[idx+1]. 
If p is not in accumulation (namely accumulation[1] > p or accumulation[end] <= p), return -1.
Bisection algorithmn is used so that the time complexity is O(log(n)) with n=length(accumulation).

MCIntegration.Dist.remove! — Method

@inline function remove!(T::Continuous, idx::Int, config)

Propose to remove old variable in [T.lower, T.lower+T.range), return proposal probability

Arguments

T: Continuous variable

idx: T.data[idx] will be updated

MCIntegration.Dist.remove! — Method

remove!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to remove the old index in [1, size]

Arguments

oldIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

MCIntegration.Dist.remove! — Method

removeFermiK!(oldK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to remove an existing Fermi K in [Kf-δK, Kf+δK)

Arguments

oldK: vector of dimension of d=2 or 3

MCIntegration.Dist.rescale — Function

function rescale(dist::AbstractVector, alpha=1.5)

rescale the dist array to avoid overreacting to atypically large number.
There are three steps:
1. dist will be first normalize to [0, 1].
2. Then the values that are close to 1.0 will not be changed much, while that close to zero will be amplified to a value controlled by alpha.
3. In the end, the rescaled dist array will be normalized to [0, 1].
Check Eq. (19) of https://arxiv.org/pdf/2009.05112.pdf for more detail

MCIntegration.Dist.shift! — Method

function shift!(T::Continuous, idx::Int, config)

Propose to shift an existing variable to a new one, both in [T.lower, T.lower+T.range), return proposal probability

Arguments

T: Continuous variable

idx: T.data[idx] will be updated

MCIntegration.Dist.shift! — Method

shift!(d::Discrete, idx::Int, config)

Propose to shift the old index in [1, size] to a new index

Arguments

oldIdx: old index ∈ [1, size]

newIdx: new index ∈ [1, size], will be modified!

size : up limit of the index

rng=GLOBAL_RNG : random number generator

MCIntegration.Dist.shift! — Method

shiftK!(oldK, newK, step, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector

MCIntegration.Dist.smooth — Function

function smooth(dist::AbstractVector, factor=6)

Smooth the distribution by averaging two nearest neighbor. The average ratio is given by 1 : factor : 1 for the elements which are not on the boundary.

MCIntegration.Dist.swap! — Method

swap!(d::Discrete, idx1::Int, idx2::Int, config)

Swap the variables idx1 and idx2

MCIntegration.Dist.train! — Method

Vegas adaptive map