Monte Carlo

MCIntegration.Configuration — Type

mutable struct Configuration

Struct that contains everything needed for MC.

There are three different componenets:

Members

Static parameters

seed: seed to initialize random numebr generator, also serves as the unique pid of the configuration

rng: a MersenneTwister random number generator, seeded by seed

para: user-defined parameter, set to nothing if not needed

var: TUPLE of variables, each variable should be derived from the abstract type Variable, see variable.jl for details). Use a tuple rather than a vector improves the performance.

integrand properties

neighbor::Vector{Tuple{Int, Int}} : vector of tuples that defines the neighboring integrands. Two neighboring integrands are directly connected in the Markov chain. e.g., [(1, 2), (2, 3)] means the integrand 1 and 2 are neighbor, and 2 and 3 are neighbor. The neighbor vector defines a undirected graph showing how the integrands are connected. Please make sure all integrands are connected. By default, we assume the N integrands are in the increase order, meaning the neighbor will be set to [(N+1, 1), (1, 2), (2, 4), ..., (N-1, N)], where the first N entries are for diagram 1, 2, ..., N and the last entry is for the normalization diagram. Only the first diagram is connected to the normalization diagram. Only highly correlated integrands are not highly correlated should be defined as neighbors. Otherwise, most of the updates between the neighboring integrands will be rejected and wasted.

dof::Vector{Vector{Int}}: degrees of freedom of each integrand, e.g., [[0, 1], [2, 3]] means the first integrand has zero var#1 and one var#2; while the second integrand has two var#1 and 3 var#2.

observable: observables that is required to calculate the integrands, will be used in the measure function call. It is either an array of any type with the common operations like +-*/^ defined.

reweight: reweight factors for each integrands. The reweight factor of the normalization diagram is assumed to be 1. Note that you don't need to explicitly add the normalization diagram.

reweight_goal: The expected distribution of visited times for each integrand after reweighting . If not set, then all factors will be initialized with one.

visited: how many times this integrand is visited by the Markov chain.

current MC state

step: the number of MC updates performed up to now

curr: the current integrand, initialize with 1

norm: the index of the normalization diagram. norm is larger than the index of any user-defined integrands

normalization: the accumulated normalization factor. Physical observable = Configuration.observable/Configuration.normalization.

absWeight: the abolute weight of the current integrand. User is responsible to initialize it after the contructor is called.

propose/accept: array to store the proposed and accepted updates for each integrands and variables. Their shapes are (number of updates X integrand number X max(integrand number, variable number). The last index will waste some memory, but the dimension is small anyway.

MCIntegration.Result — Type

struct Result{O,C}

the returned result of the MC integration.

Members

mean: mean of the MC integration

stdev: standard deviation of the MC integration

chi2: chi-square of the MC integration

neval: number of evaluations of the integrand

dof: degrees of freedom of the MC integration (number of iterations - 1)

config: configuration of the MC integration from the last iteration

iteractions: list of tuples [(data, error, Configuration), ...] from each iteration

MCIntegration.average — Function

function average(history, max=length(history))

average the history[1:max]. Return the mean, standard deviation and chi2 of the history.

Arguments

history: a list of tuples, such as [(data, error, Configuration), ...]

max: the number of data to average over

MCIntegration.create! — Method

create!(T::Continuous, idx::Int, rng=GLOBAL_RNG)

Propose to generate new (uniform) variable randomly in [T.lower, T.lower+T.range), return proposal probability

Arguments

T: Continuous variable

idx: T.data[idx] will be updated

MCIntegration.create! — Method

create!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to generate new index (uniformly) randomly in [1, size]

Arguments

newIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

MCIntegration.create! — Method

create!(K::RadialFermiK, idx::Int, rng=GLOBAL_RNG)

Propose to generate new k randomly in [0, +inf), return proposal probability k is generated uniformly on [0, K.kF-K.δk), Lorentzianly on [K.kF-K.δk,K.kF+K.δk), and exponentially on [K.kF-K.δk, +inf).

Arguments

K: k variable

idx: K.data[idx] will be updated

MCIntegration.create! — Method

create!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to generate new tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

MCIntegration.create! — Method

create!(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to generate a new pair of tau (uniformly) randomly in [0, β), return proposal probability

Arguments

T: TauPair variable

idx: T.data[idx] will be updated

MCIntegration.create! — Method

create!(K::FermiK{D}, idx::Int, rng=GLOBAL_RNG)

Propose to generate new Fermi K in [Kf-δK, Kf+δK)

Arguments

newK: vector of dimension of d=2 or 3

MCIntegration.remove! — Method

remove(T::Continuous, idx::Int, rng=GLOBAL_RNG)

Propose to remove old variable in [T.lower, T.lower+T.range), return proposal probability

Arguments

T: Continuous variable

idx: T.data[idx] will be updated

MCIntegration.remove! — Method

remove!(newIdx::Int, size::Int, rng=GLOBAL_RNG)

Propose to remove the old index in [1, size]

Arguments

oldIdx: index ∈ [1, size]

size : up limit of the index

rng=GLOBAL_RNG : random number generator

MCIntegration.remove! — Method

remove(K::RadialFermiK, idx::Int, rng=GLOBAL_RNG)

Propose to remove old k in [0, +inf), return proposal probability

Arguments

K: K variable

idx: K.data[idx] will be updated

MCIntegration.remove! — Method

remove(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to remove old tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

MCIntegration.remove! — Method

remove(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to remove an existing pair of tau in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

MCIntegration.remove! — Method

removeFermiK!(oldK, Kf=1.0, δK=0.5, rng=GLOBAL_RNG)

Propose to remove an existing Fermi K in [Kf-δK, Kf+δK)

Arguments

oldK: vector of dimension of d=2 or 3

MCIntegration.sample — Function

sample(config::Configuration, integrand::Function, measure::Function; Nblock=16, print=0, printio=stdout, save=0, saveio=nothing, timer=[])

sample the integrands, collect statistics, and return the expected values and errors.

Remarks

User may run the MC in parallel using MPI. Simply run mpiexec -n N julia userscript.jl where N is the number of workers. In this mode, only the root process returns meaningful results. All other workers return nothing, nothing. User is responsible to handle the returning results properly. If you have multiple number of mpi version, you can use "mpiexecjl" in your "~/.julia/package/MPI/###/bin" to make sure the version is correct. See https://juliaparallel.github.io/MPI.jl/stable/configuration/ for more detail.

In the MC, a normalization diagram is introduced to normalize the MC estimates of the integrands. More information can be found in the link: https://kunyuan.github.io/QuantumStatistics.jl/dev/man/important_sampling/#Important-Sampling. User don't need to explicitly specify this normalization diagram.Internally, normalization diagram will be added to each table that is related to the integrands.

Arguments

config: Configuration struct

integrand: function call to evaluate the integrand. It should accept an argument of the type Configuration, and return a weight. Internally, MC only samples the absolute value of the weight. Therefore, it is also important to define Main.abs for the weight if its type is user-defined.

measure: function call to measure. It should accept an argument of the type Configuration, then manipulate observables obs. By default, the function MCIntegration.simple_measure will be used.

neval: number of evaluations of the integrand per iteration. By default, it is set to 1e4 * length(config.dof).

niter: number of iterations. The reweight factor and the variables will be self-adapted after each iteration. By default, it is set to 10.

block: Number of blocks. Each block will be evaluated by about neval/block times. The results from the blocks will be assumed to be statistically independent, and will be used to estimate the error. The tasks will automatically distributed to multi-process in MPI mode. If the numebr of workers N is larger than block, then block will be set to be N. By default, it is set to 16.

alpha: Learning rate of the reweight factor after each iteraction. Note that alpha <=1, where alpha = 0 means no reweighting.

print: -1 to not print anything, 0 to print minimal information, >0 to print summary for every print seconds

printio: io to print the information

save: -1 to not save anything, 0 to save observables obs in the end of sampling, >0 to save observables obs for every save seconds

saveio: io to save

timer: StopWatch other than print and save.

MCIntegration.shift! — Method

shift!(T::Continuous, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing variable to a new one, both in [T.lower, T.lower+T.range), return proposal probability

Arguments

T: Continuous variable

idx: T.data[idx] will be updated

MCIntegration.shift! — Method

shift!(d::Discrete, idx::Int, config)

Propose to shift the old index in [1, size] to a new index

Arguments

oldIdx: old index ∈ [1, size]

newIdx: new index ∈ [1, size], will be modified!

size : up limit of the index

rng=GLOBAL_RNG : random number generator

MCIntegration.shift! — Method

shift!(K::RadialFermiK, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing k to a new k, both in [0, +inf), return proposal probability

Arguments

K: K variable

idx: K.data[idx] will be updated

MCIntegration.shift! — Method

shift!(T::Tau, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing tau to a new tau, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.data[idx] will be updated

MCIntegration.shift! — Method

shift!(T::TauPair, idx::Int, rng=GLOBAL_RNG)

Propose to shift an existing tau pair to a new tau pair, both in [0, β), return proposal probability

Arguments

T: Tau variable

idx: T.t[idx] will be updated

MCIntegration.shift! — Method

shiftK!(oldK, newK, step, rng=GLOBAL_RNG)

Propose to shift oldK to newK. Work for generic momentum vector

MCIntegration.summary — Function

function summary(result::Result, pick::Union{Function,AbstractVector}=obs -> real(first(obs)), name=nothing)

print the summary of the result. It will first print the configuration from the last iteration, then print the weighted average and standard deviation of the picked observable from each iteration.

Arguments

result: Result object contains the history from each iteration

pick: The pick function is used to select one of the observable to be printed. The return value of pick function must be a Number.

name: name of each picked observable. If name is not given, the index of the pick function will be used.

MCIntegration.swap! — Method

swap!(d::Discrete, idx1::Int, idx2::Int, config)

Swap the variables idx1 and idx2

MCIntegration.swap! — Method

swap!(T::Tau, idx1::Int, idx2::Int, config)

Swap the variables idx1 and idx2